New committer: Kong wei
Mr. Wei Kong has joined our project and actively work in our team since
this summer. He is leading a developer team for Escalator who are
collaborating with the Daisy4NFV installer project now.

To encourage more people to spend more efforts on Escalator in the
future with us, we should try to make something happens. I proposed to
promote Mr. Wei Kong to be one of the committers of Escalator team.

Submit this patch for Hujie according to his email.

Change-Id: I0e81f13661154054f4059e597dfd4600f127205c
Signed-off-by: Julien <zhang.jun3g@zte.com.cn>